requirements.txt
langchain
langchain-openai
langchain_community
langgraph
langsmith
OpenAI
python-dotenv
streamlit
duckduckgo-search